% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{answer}
\alias{answer}
\title{answer}
\usage{
answer(text, correct = FALSE)
}
\arguments{
\item{text}{the text do be displayed for the answer}

\item{...}{see details}
}
\description{
Construct an answer
}
\details{
Each \code{question} is made up of a number of \code{answer} calls
Each \code{answer} call require the text for the answer and a flag
to indicate whether or not it is the correct answer
}
\examples{
answer("response", correct = TRUE)
answer("response")

}
