% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{question}
\alias{question}
\title{question}
\usage{
question(text, ..., type = "standard")
}
\arguments{
\item{text}{the text do be displayed for the question}

\item{...}{see details}
}
\description{
question
}
\details{
Construct a question

Each \code{question} is made up of a number of \code{answer} calls
}
\examples{
question(
  text = "A question",
  answer("response", correct = TRUE),
  answer("response")
)
}
