% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{quiz}
\alias{quiz}
\title{quiz}
\usage{
quiz(caption = "", ...)
}
\arguments{
\item{caption}{for compatibility with learnR}

\item{...}{see details}
}
\description{
Construct quiz
}
\details{
Each \code{quiz} is made up of a number of \code{question} calls

Each \code{question} is made up of the question text and an number of
\code{answer} calls
}
\examples{
\dontrun{
# Make a simple quiz with two questions
quiz(
  caption = "",
  question(
    "What is the collective term for Windows, MacOS, Unix, *etc.*?",
    answer("Operating system", correct = TRUE),
    answer("Graphical user interface"),
    answer("Integrated development environment"),
    answer("Secondary computing software")
  ),
  question(
    "What is R?",
    answer("A high-level programming language", correct = TRUE),
    answer("An IDE for Windows and MacOS"),
    answer("Machine code for RStudio"),
    answer("Baby don't hurt me.")
  )
)
}
}
